chromium/third_party/blink/web_tests/platform/mac-mac10.10/virtual/paint-timing/external/wpt/paint-timing/fcp-only/fcp-whitespace-expected.txt

This is a testharness.js-based test.
FAIL Whitespace should not count as contentful. assert_equals: Expected first contentful paint not found.  expected 1 but got 0
Harness: the test ran to completion.
